sched: ucassist: Refactor and rename to UCLAMP scaling
Remove unnecessary code and simplify by using an array that matches
the state to its UCLAMP scale (no longer constraint). Also since
we're essentially adjusting the maximum SCHED_CAPACITY possible
based on the SCHED_CAPACITY_SCALE, let's rename the functions to
reflect that as to also be inline with scheduler nomenclature.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>