power: qcom: smb-lib: Rerun AICL after check_dash_status retries
It is possible that even though charging has not been terminated,
the AICL configuration may be wrongly configured causing charging
to effectively halt as the current drops to 0-3 mA. Rerun AICL to
refresh this configs.

Also have the FET collapse enablement to only be done once after
check_dash_status.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>